Move CompactionTaskStatus to core
