+++ java/clients/src/main/java/sleeper/clients/status/report/compaction/task/CompactionTaskStatusReporter.java
@@ -15,7 +15,7 @@
  */
 package sleeper.clients.status.report.compaction.task;
 
-import sleeper.compaction.core.task.CompactionTaskStatus;
+import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
 
 import java.io.PrintStream;
 import java.util.List;
