+++ java/clients/src/main/java/sleeper/clients/status/report/compaction/task/StandardCompactionTaskStatusReporter.java
@@ -20,8 +20,8 @@ import sleeper.clients.status.report.job.StandardProcessRunReporter;
 import sleeper.clients.util.table.TableField;
 import sleeper.clients.util.table.TableRow;
 import sleeper.clients.util.table.TableWriterFactory;
-import sleeper.compaction.core.task.CompactionTaskStatus;
 import sleeper.core.record.process.AverageRecordRate;
+import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
 
 import java.io.PrintStream;
 import java.util.List;
