+++ java/compaction/compaction-core/src/main/java/sleeper/compaction/core/task/CompactionTaskStatusesBuilder.java
@@ -16,6 +16,7 @@
 package sleeper.compaction.core.task;
 
 import sleeper.core.tracker.compaction.task.CompactionTaskFinishedStatus;
+import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
 
 import java.time.Instant;
 import java.util.Comparator;
