+++ java/compaction/compaction-core/src/test/java/sleeper/compaction/core/task/CompactionTaskStatusInPeriodTest.java
@@ -20,6 +20,7 @@ import org.junit.jupiter.api.Test;
 import sleeper.core.record.process.RecordsProcessed;
 import sleeper.core.record.process.RecordsProcessedSummary;
 import sleeper.core.tracker.compaction.task.CompactionTaskFinishedStatus;
+import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
 
 import java.time.Instant;
 import java.time.Period;
