+++ java/compaction/compaction-core/src/test/java/sleeper/compaction/core/task/CompactionTaskStatusTestData.java
@@ -18,6 +18,7 @@ package sleeper.compaction.core.task;
 
 import sleeper.core.record.process.RecordsProcessedSummary;
 import sleeper.core.tracker.compaction.task.CompactionTaskFinishedStatus;
+import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
 
 import java.time.Duration;
 import java.time.Instant;
