+++ java/compaction/compaction-core/src/test/java/sleeper/compaction/core/testutils/InMemoryCompactionTaskStatusStore.java
@@ -16,8 +16,8 @@
 
 package sleeper.compaction.core.testutils;
 
-import sleeper.compaction.core.task.CompactionTaskStatus;
 import sleeper.compaction.core.task.CompactionTaskStatusStore;
+import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
 
 import java.time.Instant;
 import java.util.LinkedHashMap;
