+++ java/compaction/compaction-status-store/src/test/java/sleeper/compaction/status/store/task/QueryAllCompactionTasksIT.java
@@ -17,8 +17,8 @@ package sleeper.compaction.status.store.task;
 
 import org.junit.jupiter.api.Test;
 
-import sleeper.compaction.core.task.CompactionTaskStatus;
 import sleeper.compaction.status.store.testutils.DynamoDBCompactionTaskStatusStoreTestBase;
+import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
 
 import java.time.Instant;
