+++ java/core/src/main/java/sleeper/core/tracker/compaction/task/CompactionTaskStatus.java
@@ -14,13 +14,11 @@
  * limitations under the License.
  */
 
-package sleeper.compaction.core.task;
+package sleeper.core.tracker.compaction.task;
 
 import sleeper.core.record.process.status.ProcessFinishedStatus;
 import sleeper.core.record.process.status.ProcessRun;
 import sleeper.core.record.process.status.TimeWindowQuery;
-import sleeper.core.tracker.compaction.task.CompactionTaskFinishedStatus;
-import sleeper.core.tracker.compaction.task.CompactionTaskStartedStatus;
 
 import java.time.Duration;
 import java.time.Instant;
