+++ java/system-test/system-test-dsl/src/test/java/sleeper/systemtest/dsl/testutil/drivers/InMemoryReports.java
@@ -15,8 +15,8 @@
  */
 package sleeper.systemtest.dsl.testutil.drivers;
 
-import sleeper.compaction.core.task.CompactionTaskStatus;
 import sleeper.core.tracker.compaction.job.query.CompactionJobStatus;
+import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
 import sleeper.ingest.core.job.status.IngestJobStatus;
 import sleeper.systemtest.dsl.instance.SystemTestInstanceContext;
 import sleeper.systemtest.dsl.reporting.CompactionReportsDriver;
